+++ diffusion/timestep_sampler.py
@@ -123,13 +123,13 @@ class LossSecondMomentResampler(LossAwareSampler):
         self.history_per_term = history_per_term
         self.uniform_prob = uniform_prob
         self._loss_history = np.zeros(
-            [diffusion.num_timesteps, history_per_term], dtype=np.float64
+            [diffusion.num_timesteps, history_per_term], dtype=np.float32
         )
-        self._loss_counts = np.zeros([diffusion.num_timesteps], dtype=np.int)
+        self._loss_counts = np.zeros([diffusion.num_timesteps], dtype=int)
 
     def weights(self):
         if not self._warmed_up():
-            return np.ones([self.diffusion.num_timesteps], dtype=np.float64)
+            return np.ones([self.diffusion.num_timesteps], dtype=np.float32)
         weights = np.sqrt(np.mean(self._loss_history ** 2, axis=-1))
         weights /= np.sum(weights)
         weights *= 1 - self.uniform_prob
